add apps sidebar tab
